+++ src/rays/rays.rs
@@ -14,7 +14,7 @@ impl Ray {
         Ray { origin, direction }
     }
 
-    pub fn position(&self, t: f32) -> Tuple {
+    pub fn position(&self, t: f64) -> Tuple {
         self.origin + self.direction * t
     }
